rename to digital_ocean_utils
