+++ src/cpdb/helpers.py
@@ -1,7 +1,7 @@
 import pandas as pd
 from dotenv import load_dotenv
 import geopandas as gpd
-from src.digital_ocean_client import DigitalOceanClient
+from src.digital_ocean_utils import DigitalOceanClient
 
 load_dotenv()
